Remove pinia
I added pinia during scaffolding, but it turns out it's simple enough that we don't need that.

In the original BackStage + Wings era, the UI heavily relied on stores to manage data. But the new architecture is so much more simple with it only having the vote element, it's actually kinda beautiful.

Anyway good night